no need for root permissions anymore if you have installed all packages

+++ install-ubuntu.sh
@@ -18,56 +18,81 @@ fi
 
 function installPackages {
 	printf "\nMaking sure essentials are installed...\n"
-	sudo apt-get -qq update # quiet-mode
-	# no checking here because "apt-get update" can produce non-critical errors
-	# e.g. non-critical repository not found
 
-	sudo apt-get install build-essential
+	pkg=( g++ make automake libtool xutils-dev m4 libreadline-dev libgsl-dev \
+	libglu1-mesa-dev libgl1-mesa-dev freeglut3-dev libopenscenegraph-dev libqt4-dev libqt4-opengl \
+	libqt4-opengl-dev qt4-qmake libqt4-qt3support gnuplot gnuplot-x11 libncurses5-dev )
 
-	# check if packages "build-essentials" are installed correctly
-	chExitStatus
+	toInstall=()
 
-	printf "\nInstalling necessary packages for compiling...\n"
-	sleep 2
-	sudo apt-get install g++ make automake libtool xutils-dev m4 libreadline-dev libgsl0-dev \
-	libglu-dev libgl1-mesa-dev freeglut3-dev libopenscenegraph-dev libqt4-dev libqt4-opengl \
-	libqt4-opengl-dev qt4-qmake libqt4-qt3support gnuplot gnuplot-x11 libncurses5-dev
+	for name in ${pkg[*]}; do
+		printf "Checking if $name is installed..."
+		if ! dpkg -s $name > /dev/null; then
+			toInstall+=( $name )
+			printf " - No\n"
+			continue
+		fi
+		printf " - Ok\n"
+	done
 
-	# check if all packages were installed correctly
-	chExitStatus
+	if [[ -z '$toInstall' ]]; then
+		printf "\nInstalling necessary packages for compiling...\n"
+		sudo apt-get -qq update # quiet-mode
 
-	printf "\nAll packages necessary for compiling are now installed.\n"
+		sudo apt-get install ${toInstall[*]}
+
+		chExitStatus
+
+		printf "\nAll packages necessary for compiling are installed.\n"
+	fi	
 }
 
 function makeProgram {
 	# call with location of files as first argument!
 
+	# replace installation location for easier install w/o sudo
+	#sed -i "s/{1:-\/usr\/local}/{1:-\/home\/${USER}\/Documents}/g" createMakefile.conf.sh
+	#sed -i "s/Please use either \/usr, \/usr\/local  or you home directory/Please use either \/home\/${USER}\/Documents or your home directory/g" createMakefile.conf.sh
+	
+	# taken directly form the createMakefile.conf.sh-File
+	prefix="/home/$USER"
+	export PATH=$prefix/bin:$PATH
+
+	chExitStatus
+
+	# taken directly form createMakefile.conf.sh
+	echo -e "# configuration file for lpzrobots (automatically generated!)\n\
+	# Where to install the simulator and utils"  > Makefile.conf
+	echo "PREFIX=$prefix" >> Makefile.conf
+	echo -e "\n# user or developement installation\n\
+	# Options:\n\
+	#  DEVEL: only install the utility functions,\n\
+	#   which is useful for development on the simulator\n\
+	#  USER: install also the ode_robots and selforg libaries and include files\n\
+	#   this is recommended for users" >> Makefile.conf
+	echo "TYPE=DEVEL" >> Makefile.conf
+
+	echo "// Automatically generated file! Use make conf in lpzrobots." > ode_robots/install_prefix.conf
+	echo "#define PREFIX \"$prefix\"" >> ode_robots/install_prefix.conf
+
+
 	printf "\nStarting the make-process.\n"	
 	sleep 1
-
-	# make symlink (otherwise there will be errors)
-	sudo ln -sf ${1}/LpzRobots/lpzrobots-master/opende/ode/src/.libs/libode_dbl.so.1 /lib/libode_dbl.so.1
 
 	# start timer
 	START=$(date +%s.%N)
 
-	make
-
-	wait
-
-	# check if make process was successful
-	chExitStatus
-
 	# start make-process
-	sudo make all
-
-	wait
+	make all
 
 	# check if build process was successful
 	chExitStatus
 
 	# end timer
 	END=$(date +%s.%N)
+
+	# make symlink (otherwise there will be errors)
+	#sudo ln -sf ${1}/LpzRobots/lpzrobots-master/opende/ode/src/.libs/libode_dbl.so.1 /lib/libode_dbl.so.1
 
 	# calculate difference
 	DIFF=$(echo "($END - $START)" | bc)
@@ -82,24 +107,24 @@ function getFiles {
 	# move into the directory the user has specified
 	cd $1
 
-	# make directory LpzRobots (exits with error if directory already exists)
-	mkdir LpzRobots
+	# make directory LpzRobots
+	if [[ ! -d LpzRobots ]]; then mkdir LpzRobots; fi
 
 	# move into the newly created directory
 	cd LpzRobots
 
-	printf "Downloading files from github...\n\n"
+	printf "Getting files from github...\n"
 	wget --quiet https://github.com/georgmartius/lpzrobots/archive/master.zip
 
 	chExitStatus
 
-	printf "Unzipping content...\n\n"
+	printf "Unzipping content...\n"
 	unzip -q master.zip
 
 	# check if directory exists
 	if [[ ! -e lpzrobots-master ]]; then
 		printf "\nSorry, something went wrong downloading or unzipping the files \
-	('lpzrobots-master'-directory does not exist). - ABORT\n"
+	('lpzrobots-master'-directory does not exist). - Exiting\n"
 		exit 1
 	fi
 
@@ -137,10 +162,10 @@ function testInstall {
 
 function cleanUp {
 	printf "\nCleaning up...\n"
-	rm ${1}/LpzRobots/master.zip
+	if [[ -e ${1}/LpzRobots/master.zip ]];then rm ${1}/LpzRobots/master.zip; fi
 }
 
-trap "printf '\nExiting...\n'; exit 0" INT TERM
+trap "printf '\nExiting...\n'; exit 1" INT TERM
 
 
 ######	Initial setup  ######
@@ -149,47 +174,32 @@ trap "printf '\nExiting...\n'; exit 0" INT TERM
 clear
 
 # display some information about the program to the user
-printf "\nThis script will download and install the lpzrobots-environment for you.\n\n"
-printf "It has been tested on Ubuntu 14.04. Default values are in '[]' and/or capitalized, \
-just press ENTER to use them.\n\n"
+printf "\nThis script will download and install the lpzrobots-environment for you.\n"
+printf "It has been tested on Ubuntu 14.04 and 16.04. Default values are in '[]' and/or capitalized, \
+just press ENTER to use them.\n"
 printf "THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND.\n"
 
 # ask the user in which directory the files should go
-while true; do
-printf "\nIn which directory should the program be downloaded and compiled? ('/LpzRobots' \
-will be added to the end of the path, don't use '~') [/home/${USER}/Downloads]\n"
+while :
+do
+	defLoc="/home/${USER}"
+	printf "\nIn which directory should the program be downloaded and compiled? ('/LpzRobots' \
+	will be added to the end of the path, don't use '~') [$defLoc]\n"
 
-# read the answer from the command line
-read location
+	# read the answer from the command line
+	read location
 
-# if location is empty take default value otherwise just keep the input
-if [[ -z $location ]]; then
-	location="/home/${USER}/Downloads"
-fi
+	# if location is empty take default value otherwise just keep the input
+	if [[ -z $location ]]; then	location=$defLoc; fi
 
-# check if $location exists, is a directory and if we have read/write access
-if [[ ! -d $location || ! -w $location ]]; then
-	printf "Directory does not exist or you do not have read/write-access! \nChoose another one.\n"
-	continue
-fi
-
-# ask user for confirmation
-printf "\nThe lpzrobots-files will now be downloaded into the directory ${location}/LpzRobots.\nIs this OK? [n/Y]\n"
-
-read ans
-
-# check answer
-if [[ $ans == "n" ]]; then
-	echo
-	continue
-elif [[ $ans == "" || $ans == "Y" ]]; then
-	printf "OK, continuing.\n"
-	break
-else
-	printf "Sorry, did not catch that!\n\n"
-	continue
-fi
-
+	# check if $location exists, is a directory and if we have read/write access
+	if [[ ! -d $location || ! -w $location ]]; then
+		printf "Directory does not exist or you do not have read/write-access! \nChoose another one.\n"
+		continue
+	else
+		printf "\nThe lpzrobots-files will now be downloaded into the directory ${location}/LpzRobots.\n"
+		break
+	fi
 done
 
 # set up the trap for the clean up process
@@ -207,46 +217,17 @@ installPackages
 
 ######  Build/compile program  ######
 
-# ask user if they want to continue with the compile-process
-while true; do
-printf "The next command will compile the program, it will take a long time to finish. \nDo you want to continue? (type no/yes)\n"
-
-read ans
-
-# again check answer
-if [[ $ans == "no" ]]; then
-	echo "Exiting."
-	exit 0
-elif [[ $ans == "yes" ]]; then
-	printf "\nOK, continuing.\n\n"
-	makeProgram $location
-	break
-else
-	printf "Sorry, did not catch that!\n\n"
-	continue
-fi
-done
+makeProgram $location
 
 
 ######  Check installation  ######
 
-while true; do
 printf "\nDo you want to test if the program is installed correctly? [n/Y]"
 
 read ans
 
 # again check answer
-if [[ $ans == "n" ]]; then
-	break
-elif [[ $ans == "Y" || $ans == "y" || -z $ans ]]; then
-	testInstall $location
-	break
-else
-	printf "Sorry, did not catch that!\n\n"
-	continue
-fi
-done
-
+if [[ $ans == "Y" || $ans == "y" || -z $ans ]]; then testInstall $location; fi
 
 printf "\nThat's it! Have a nice day.\n"
